+++ slideflow/project.py
@@ -7,7 +7,6 @@ import json
 import multiprocessing
 import numpy as np
 import os
-import sys
 import shutil
 import pickle
 import pandas as pd
